+++ src/main/java/de/skuzzle/ghpromexporter/appmetrics/AppMetrics.java
@@ -1,5 +1,7 @@
 package de.skuzzle.ghpromexporter.appmetrics;
 
+import io.micrometer.core.instrument.Counter;
+import io.micrometer.core.instrument.DistributionSummary;
 import io.micrometer.core.instrument.Metrics;
 import io.micrometer.core.instrument.Timer;
 
@@ -12,8 +14,33 @@ public final class AppMetrics {
     private static final String NAMESPACE = "ghp_";
 
     private static final Timer scrapeDuration = Metrics.timer(NAMESPACE + "repository_scrape_duration");
+    private static final DistributionSummary registeredScrapers = Metrics.summary(NAMESPACE + "registered_scrapers");
+    private static final Counter scrapeFailures = Metrics.counter(NAMESPACE + "scrape_failures");
+    private static final Counter abuses = Metrics.counter(NAMESPACE + "abuses");
+    private static final Counter apiCalls = Metrics.counter(NAMESPACE + "api_calls");
+    private static final Counter rateLimitHits = Metrics.counter(NAMESPACE + "rate_limit_hits");
 
     public static Timer scrapeDuration() {
         return scrapeDuration;
+    }
+
+    public static DistributionSummary registeredScrapers() {
+        return registeredScrapers;
+    }
+
+    public static Counter scrapeFailures() {
+        return scrapeFailures;
+    }
+
+    public static Counter abuses() {
+        return abuses;
+    }
+
+    public static Counter apiCalls() {
+        return apiCalls;
+    }
+
+    public static Counter rateLimitHits() {
+        return rateLimitHits;
     }
 }