+++ src/main/java/de/skuzzle/ghpromexporter/github/AuthenticationProvider.java
@@ -1,8 +1,0 @@
-package de.skuzzle.ghpromexporter.github;
-
-import org.springframework.http.server.reactive.ServerHttpRequest;
-
-public interface AuthenticationProvider {
-
-    GitHubAuthentication authenticateRequest(ServerHttpRequest request);
-}