+++ src/main/java/de/skuzzle/ghpromexporter/github/GitHubConfiguration.java
@@ -3,7 +3,6 @@ package de.skuzzle.ghpromexporter.github;
 import javax.annotation.PostConstruct;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration(proxyBeanMethods = false)
@@ -21,8 +20,4 @@ public class GitHubConfiguration {
         GitHubFactory.CACHED_GITHUBS = properties.cache().build();
     }
 
-    @Bean
-    AuthenticationProvider authenticationProvider() {
-        return GitHubAuthentication::fromRequest;
-    }
 }