+++ src/main/java/de/skuzzle/ghpromexporter/scrape/AsynchronousScrapeService.java
@@ -6,11 +6,15 @@ import org.springframework.cloud.sleuth.Span;
 import org.springframework.cloud.sleuth.Tracer;
 import org.springframework.scheduling.annotation.Scheduled;
 
+import de.skuzzle.ghpromexporter.appmetrics.AppMetrics;
 import de.skuzzle.ghpromexporter.github.GitHubAuthentication;
 import de.skuzzle.ghpromexporter.scrape.RegistrationRepository.RegisteredScraper;
 import reactor.core.publisher.Mono;
-import reactor.core.scheduler.Schedulers;
 
+/**
+ *
+ * @author Simon Taddiken
+ */
 public class AsynchronousScrapeService {
 
     private static final Logger log = LoggerFactory.getLogger(AsynchronousScrapeService.class);
@@ -28,14 +32,16 @@ public class AsynchronousScrapeService {
 
     public Mono<RepositoryMetrics> scrapeReactive(GitHubAuthentication authentication,
             ScrapeRepositoryRequest request) {
-        final RegisteredScraper registeredScraper = new RegisteredScraper(authentication, request);
+        final RegisteredScraper scrapeTarget = new RegisteredScraper(authentication, request);
 
-        return registrationRepository.getExistingOrLoad(registeredScraper, () -> {
-            final RepositoryMetrics repositoryMetrics = registeredScraper.scrapeWith(scrapeRepositoryService);
-            log.info("Cache miss for {}. Scraped fresh metrics now in {}ms", registeredScraper,
-                    repositoryMetrics.scrapeDuration());
-            return repositoryMetrics;
-        });
+        return registrationRepository
+                .getExistingOrLoad(scrapeTarget, scraper -> {
+                    final RepositoryMetrics repositoryMetrics = scraper.scrapeWith(scrapeRepositoryService);
+                    log.info("Cache miss for {}. Scraped fresh metrics now in {}ms", scraper,
+                            repositoryMetrics.scrapeDuration());
+                    return repositoryMetrics;
+                })
+                .doOnError(error -> AppMetrics.scrapeFailures().increment());
     }
 
     @Scheduled(
@@ -47,28 +53,31 @@ public class AsynchronousScrapeService {
             // return early to improve INFO logging
             return;
         }
+        AppMetrics.registeredScrapers().record(registrationRepository.estimatedCount());
 
         final Span newSpan = tracer.nextSpan().name("scheduledScrape");
         try (var ws = tracer.withSpan(newSpan.start())) {
             registrationRepository.registeredScrapers()
-                    .doOnNext(scraper -> scrapeAndUpdateCache(newSpan, scraper))
+                    .doOnNext(scrapeTarget -> scrapeAndUpdateCache(newSpan, scrapeTarget))
                     .doOnTerminate(() -> log.info("Updated cached metrics for all registered scrapers"))
-                    .subscribeOn(Schedulers.boundedElastic())
-                    .subscribe();
+                    .blockLast();
         } finally {
             newSpan.end();
         }
     }
 
-    private void scrapeAndUpdateCache(Span parentSpan, RegisteredScraper scraper) {
+    private void scrapeAndUpdateCache(Span parentSpan, RegisteredScraper scrapeTarget) {
         final Span nextSpan = tracer.nextSpan(parentSpan).name("scrapeSingleRepo");
         try (var ws = tracer.withSpan(nextSpan.start())) {
-            final RepositoryMetrics repositoryMetrics = scraper.scrapeWith(scrapeRepositoryService);
-            registrationRepository.updateRegistration(scraper, repositoryMetrics);
-            log.info("Asynschronously updated metrics for: {} in {}ms", scraper, repositoryMetrics.scrapeDuration());
+            final RepositoryMetrics repositoryMetrics = scrapeTarget.scrapeWith(scrapeRepositoryService);
+            registrationRepository.updateRegistration(scrapeTarget, repositoryMetrics);
+            log.info("Asynschronously updated metrics for: {} in {}ms", scrapeTarget,
+                    repositoryMetrics.scrapeDuration());
         } catch (final Exception e) {
-            registrationRepository.deleteRegistration(scraper);
-            log.error("Scrape using '{}' threw exception. Will be removed from cache of active scrapers", scraper, e);
+            registrationRepository.deleteRegistration(scrapeTarget);
+            AppMetrics.scrapeFailures().increment();
+            log.error("Scrape using '{}' threw exception. Will be removed from cache of active scrapers", scrapeTarget,
+                    e);
         } finally {
             nextSpan.end();
         }